added parsing

+++ World.cpp
@@ -2,6 +2,6 @@
 
 #include "World.h"
 
-World::World(size_t max_turns) : m_max_turns(max_turns), m_current_turn(0) {}
+World::World(int max_turns) : m_max_turns(max_turns), m_current_turn(0) {}
